+++ server/server.js
@@ -4,13 +4,16 @@
 // all routing.
 
 var express = require('express');
-var mongoose = require('mongoose');
-var mongodb =require('../keys.js').mongodb;
+// var mongoose = require('mongoose');
+
+// Initialize the SQLite databases
+var sql =require('../keys.js').sqlite3;
+// var db = Bookshelf.initialize(sql);
 
 // Defer to the locally hosted mongodb if a process
 // variable is not defined.
-var dbURL = mongodb || 'mongodb://localhost/nite-out';
-mongoose.connect(dbURL);
+// var dbURL = mongodb || 'mongodb://localhost/nite-out';
+// mongoose.connect(dbURL);
 
 var app = express();
